services/ingestion/requirements.txt
kafka-python==2.0.2
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
influxdb-client==1.43.0
python-dotenv==1.0.1
